Add files via upload
The PyCaret package is very useful and has many advantages. It automatically handles preprocessing, builds models, and compares them to find the most suitable one for the dataset, highlighting the best choice. This allows us to further proceed with the model and fine-tune the hyperparameters. It is a very useful library, especially if you are working with small datasets.